app/requirements.txt
boto3
mypy-boto3-sqs
pyyaml

moto
pytest
pytest-cov